standardize code and utilize types
