--- client/src/store/actions/types.ts
@@ -1,3 +1,0 @@
-export const FETCH_USER = 'fetch_user';
-export const FETCH_SNIPPETS = 'fetch_snippets';
-export const AUTHORIZE_USER = 'authorize_user';
