+++ client/src/store/reducers/authReducer.ts
@@ -1,16 +1,17 @@
-import { FETCH_USER, AUTHORIZE_USER } from '../actions/types';
+import {
+    FETCH_USER,
+    AUTHORIZE_USER
+} from '../types';
 
-export default function(state = {}, action: any) {
+const authReducer = (state = {}, action: any) => {
     switch (action.type) {
         case FETCH_USER:
             return action.payload || false;
         case AUTHORIZE_USER:
-            // return {
-            //     ...state,
-            //     user: action.user || null,
-            // };
             return action.user || null;
         default:
             return state;
     }
 };
+
+export default authReducer;
