+++ client/src/store/reducers/snippetReducer.ts
@@ -1,10 +1,12 @@
-import { FETCH_SNIPPETS } from '../actions/types';
+import { FETCH_SNIPPETS } from '../types'
 
-export default function(state = [], action: any) {
+const snippetReducer = (state = [], action: any) => {
     switch (action.type) {
         case FETCH_SNIPPETS:
             return action.payload;
         default:
             return state;
     }
-}
+};
+
+export default snippetReducer;
