--- client/src/store/stateTypes.tsx
@@ -1,13 +1,0 @@
-
-
-export interface User {
-    email: string;
-    firstName: string;
-    googleId: string;
-    picture: string;
-    lastName: string;
-}
-
-export interface State {
-    user: User | null;
-}
